Provide unsafe type class instances for Encoder[Char] and Decoder[Char]
The instances are unsafe because they map to/from `AttributeValue`s as
strings.  In the decoding case, we take the first character of the
string as the value.  If the string is empty or null, this will throw at
runtime.

Codec tests updated to prove the round-trip property of
encoding/decoding.